Fixing problem with killing an no existent docker image

+++ bin/deploy.sh
@@ -5,8 +5,11 @@ echo Running $1 on port $2 on $3
 
 ssh vagrant@$3 '
 	(echo Removing currently running images &&
+	if [ ! -z "tictactoe$2" ]
+	then
 	docker kill tictactoe$2 &&
-	docker rm  tictactoe$2 &&
+	docker rm  tictactoe$2
+	fi &&
 	echo Pulling newest docker image &&
 	docker pull fanneyyy/tictactoe &&
 	echo Running newest docker image &&
